requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
beautifulsoup4==4.12.3
    # via -r requirements.in
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
pillow==10.4.0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
soupsieve==2.5
    # via beautifulsoup4
urllib3==2.2.2
    # via requests
